[WIP][DRAFT] Trying to implement pulp-server secret template
This is just a draft!
This commit does not compile and is not tested at all!
I tried to create a new field on CRD of type map[string]interface{}
so that we could kind of reproduce the same configuration from
"default_settings" dictionary, used to populate pulp-server secret in
original pulp-operator. The problem is that this is something not allowed.
I'm keeping this as a commit just for future reference.